## 0.0.3 - SNAPSHOT - 2017-12-15 ### Added - 实现自动生成REST API文档（使用SWAGGER2框架）。 - 实现用户登录接口。 - 添加系统工程表。 - 添加部署设计表. - 实现工程的保存和删除、更新、查询和管理员查询接口。 - 实现部署设计表的增加、删除、修改、查询、管理员查询等功能。 - 实现部署设计绑定设备和组件及部署路径 - 创建NotificationMessage类，统一处理错误信息。 ### Changed - 修改构建返回结果对象的实现逻辑。 - 修改设备和组件表的表结构，除去用户外键。 - 修改现有设备和组件的操作逻辑。 - 修改现有表结构，添加一些约束条件。 - 重构用户、角色部分接口的实现，分离角色操作。 ### Fixed - 修复返回结果对象中，loginUser可能为null引起的空指针的问题。 - 修复请求出现错误时也能不能正确获取用户名称的问题。 - 修复了更新设备信息时，设备创建时间会被修改的问题。 - 修复了PATCH和PUT、DELETE请求不能响应跨域请求的问题。 - 修复了出现401错误时，不能正确获取用户名的问题。 - 修复了不同用户之间工程名称不能相同的问题。 ### Removed
